requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4

# Data Visualization
streamlit==1.28.2
plotly==5.17.0
folium==0.15.1
streamlit-folium==0.25.1

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2
xgboost==2.0.3
lightgbm==4.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Web Framework & API
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
aiohttp==3.9.1

# Data Collection & Processing
tqdm==4.66.1
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.2

# Geospatial
geopandas==0.14.1
shapely==2.0.2
pyproj==3.6.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7

# Caching
redis==5.0.1
cachetools==5.3.2

# Configuration
pydantic==2.5.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Production
gunicorn==21.2.0
supervisor==4.2.5

# Additional ML Libraries
tensorflow==2.15.0
torch==2.1.1
transformers==4.36.2

# Time Series Analysis
prophet==1.1.5
statsmodels==0.14.0

# Advanced Analytics
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# Cloud & Deployment
boto3==1.34.0
google-cloud-storage==2.10.0
azure-storage-blob==12.19.0

# Security
cryptography==41.0.8
passlib==1.7.4
python-jose==3.3.0

# Performance
numba==0.58.1
cython==3.0.8
